+++ lib/src/code_generators/v3/swagger_models_generator_v3.dart
@@ -96,7 +96,23 @@ class SwaggerModelsGeneratorV3 extends SwaggerModelsGenerator {
           return;
         }
 
-        final properties = map['properties'] as Map<String, dynamic>;
+        Map<String, dynamic> properties;
+
+        if (mapMap.containsKey('allOf')) {
+          final allOf = mapMap['allOf'] as List<dynamic>;
+          var propertiesContainer = allOf.firstWhere(
+              (e) => (e as Map<String, dynamic>).containsKey('properties'),
+              orElse: () => null) as Map<String, dynamic>;
+
+          if (propertiesContainer != null) {
+            properties =
+                propertiesContainer['properties'] as Map<String, dynamic>;
+          } else {
+            properties = map['properties'] as Map<String, dynamic>;
+          }
+        } else {
+          properties = map['properties'] as Map<String, dynamic>;
+        }
 
         if (properties == null) {
           return;