+++ test/generator_tests/requests_generator_test.dart
@@ -289,7 +289,8 @@ void main() {
           '/test/items',
           'get',
           <ResponseOverrideValueMap>[],
-          []);
+          [],
+          {});
 
       expect(result, equals('TestItem'));
     });
@@ -535,7 +536,8 @@ void main() {
           ]),
           GeneratorOptions(),
           [],
-          []);
+          [],
+          {});
 
       expect(result, contains('Future<chopper.Response> testPathGet();'));
     });
@@ -557,7 +559,8 @@ void main() {
           ]),
           GeneratorOptions(),
           [],
-          []);
+          [],
+          {});
 
       expect(result, contains('Future<chopper.Response> _testPathGet'));
     });
@@ -579,7 +582,8 @@ void main() {
           ]),
           GeneratorOptions(),
           [],
-          []);
+          [],
+          {});
 
       expect(result, contains('Future<chopper.Response> _testPathGet'));
     });
@@ -601,7 +605,8 @@ void main() {
           ]),
           GeneratorOptions(),
           [],
-          []);
+          [],
+          {});
 
       expect(result, contains('Future<chopper.Response> _testPathGet'));
     });