+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.6.0
+
+* Added new build option use_inheritance to enable/disable inheritance support for v3
+* Fixed issue with basic types (String, int...) used as definitions
+
 # 1.5.1
 
 * Added 'export file_name.swagger.enums.dart' to main generated file